+++ src/SignatureDecipher.php
@@ -54,31 +54,6 @@ class SignatureDecipher
 	}
 
 	/**
-	 * download the raw player script with cache
-	 *
-	 * @deprecated since version 0.3, to be removed in 0.4. Use SignatureDecipher::downloadRawPlayerScript() instead
-	 *
-	 * @param string $videoID
-	 * @return string returns the playerID
-	 */
-	public static function downloadPlayerScript($videoID)
-	{
-		@trigger_error(__METHOD__ . ' is deprecated since version 0.3, to be removed in 0.4. Use SignatureDecipher::downloadRawPlayerScript() instead', E_USER_DEPRECATED);
-
-		$player_info = static::getPlayerInfoByVideoId($videoID);
-
-		$playerID = $player_info[0];
-		$playerURL = $player_info[1];
-
-		if(!file_exists('cache/playerscript_' . $playerID)) {
-			$decipherScript = self::downloadRawPlayerScript($playerURL);
-			file_put_contents('cache/playerscript_' . $playerID, $decipherScript);
-		}
-
-		return $playerID;
-	}
-
-	/**
 	 * decipher a signature with a raw player script
 	 *
 	 * @param string $decipherScript
@@ -157,36 +132,6 @@ class SignatureDecipher
 
 		//Return signature
 		return $decipheredSignature;
-	}
-
-	/**
-	 * decipher a signature
-	 *
-	 * @deprecated since version 0.3, to be removed in 0.4. Use SignatureDecipher::decipherSignatureWithRawPlayerScript() instead
-	 *
-	 * @param string $playerID
-	 * @param string $signature
-	 * @return string returns the decipherd signature
-	 */
-	public static function decipherSignature($playerID, $signature)
-	{
-		@trigger_error(__METHOD__ . ' is deprecated since version 0.3, to be removed in 0.4. Use SignatureDecipher::decipherSignatureWithRawPlayerScript() instead', E_USER_DEPRECATED);
-
-		if(!$playerID) return;
-
-		if(file_exists("cache/playerscript_$playerID"))
-		{
-			$decipherScript = file_get_contents("cache/playerscript_$playerID");
-		}
-		else
-		{
-			die("\n==== Player script was not found for id: $playerID ====");
-		}
-
-		return static::decipherSignatureWithRawPlayerScript(
-			$decipherScript,
-			$signature
-		);
 	}
 
 	private static function executeSignaturePattern($patterns, $deciphers, $signature)